requirements.txt
beautifulsoup4==4.12.2
filetype==1.2.0
pandas==2.1.2
python_dateutil==2.8.2
Requests==2.31.0
selenium==4.14.0
webdriver_manager==4.0.1
